Port src/marks/{delaunay,hexgrid} from JS to TS (stage 1)
Combine each .js + .d.ts pair into a single .ts file, inlining the
hand-written types. No behaviour changes.
